requirements.txt
asgiref==3.4.1
awsebcli==3.20.10
backports.entry-points-selectable==1.1.1
botocore==1.31.62
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.3
Deprecated==1.2.13
distlib==0.3.4
Django==4.0
filelock==3.4.0
future==0.16.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
mysqlclient==2.2.0
networkx==2.6.3
numpy==1.26.0
pathspec==0.10.1
platformdirs==2.4.0
python-dateutil==2.8.2
python-tsp==0.2.1
PyYAML>=5.3.1
requests>=2.31
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
tabulate==0.8.9
termcolor==1.1.0
tsplib95==0.7.1
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.10.0
wcwidth==0.1.9
whitenoise==5.3.0
wrapt==1.13.3
